com.unity.xr.openvr.standalone@1.0.0 # Changelog All notable changes to this package will be documented in this file.
The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [1.0.0] - 2018-04-20

Rename to com.unity.xr.openvr.standalone

## [1.0.0] - 2018-04-19

First official version of OpenVR package.

## [0.0.3] - 2018-03-29

Add Documentation.meta file

## [0.0.1] - 2018-03-01

### This is the first release of the OpenVR XR Package.

*Short description of this release*
This is the first test iteration of the OpenVR XR Package using the Package Manager process. This package is a replica of the builds.zip hosted at:
https://ono.unity3d.com/unity-extra/vr-openvr-external#trunk

We are bringing over all libraries from the builds.zip to be included and distributed through the package manager.